I believe I fixed the issue Pete raised
With `--pantheonV2` enabled we weren't fixing up the source,
but we are now. It looks like there may be some extra ifdefs that don't
need to be there, but they're empty so it doesn't seem to be affecting
anything. I've checked output on the getting started section and it
looks correct to me.